Rename project from Lapwing to Yearbird; major update to version 0.3
- Renamed all references from Lapwing to Yearbird throughout codebase
- Added Help/User Guide menu with HTML user guide (guide/guide_Yearbird.html)
- Added dark theme, palette, and stylesheet support
- Added photo management features
- Added choropleth map support (US states, US counties, world countries)
- Added .gitignore; removed stale .pyc files, old zips, and Lapwing icons
- Removed old user_guide/ and eBirdMain.spec; cleaned up obsolete data files

+++ code_Compare.py
@@ -287,8 +287,8 @@ class Compare(QMdiSubWindow, form_Compare.Ui_frmCompare):
     def scaleMe(self):
                
         scaleFactor = self.mdiParent.scaleFactor
-        windowWidth =  900  * scaleFactor
-        windowHeight = 500 * scaleFactor            
+        windowWidth =  int(900  * scaleFactor)
+        windowHeight = int(500 * scaleFactor)
         self.resize(windowWidth, windowHeight)
         
         fontSize = self.mdiParent.fontSize
